- purge unused codes

+++ app/components/Dashboard/field_survey/field_survey_logs.js
@@ -1,13 +1,13 @@
+import moment from "moment";
+import { Icon } from 'native-base';
 import React, { Component } from 'react';
-import { View, Text, ScrollView, TouchableOpacity, Dimensions, Alert } from 'react-native';
-import { field_survey_styles } from '../../../assets/styles/field_survey_styles'
-import { defaults } from '../../../assets/styles/default_styles'
-import { DataTable } from 'react-native-paper'
-import { Icon } from 'native-base'
-import Storage from '../../utils/storage'
-import { NavigationEvents } from 'react-navigation'
-import moment from "moment"
-import Notification from '../../utils/alert_notification'
+import { Alert, Dimensions, ScrollView, Text, TouchableOpacity, View } from 'react-native';
+import { DataTable } from 'react-native-paper';
+import { NavigationEvents } from 'react-navigation';
+import { defaults } from '../../../assets/styles/default_styles';
+import { field_survey_styles } from '../../../assets/styles/field_survey_styles';
+import Notification from '../../utils/alert_notification';
+import Storage from '../../utils/storage';
 
 export default class FieldSurveyLogs extends Component {
   constructor(props) {
@@ -133,34 +133,36 @@ export default class FieldSurveyLogs extends Component {
         let field_logs = [];
         let to_local_data = [];
         let counter = 0
-        for (const [index, value] of responseJson.entries()) {
-          let format_date_time = this.formatDateTime(date = value.date);
+        if (responseJson.length != 0) {
+          for (const [index, value] of responseJson.entries()) {
+            let format_date_time = this.formatDateTime(date = value.date);
+            field_logs.push(<DataTable.Row style={{ width: 500 }}>
+              <DataTable.Cell style={{ marginRight: -90 }}>{format_date_time["text_format_timestamp"]}</DataTable.Cell>
+              <DataTable.Cell style={{ marginRight: 10 }}>Field Survey Report {value.date.split(' ')[0]}</DataTable.Cell>
+              <DataTable.Cell style={{ marginRight: -190 }}>
+                <Icon name="md-create" style={{ color: "blue" }} onPress={() => this.updateLog(value)}></Icon><Text>   </Text>
+                <Icon name="ios-trash" style={{ color: "red" }} onPress={() => this.removeConfirmation(value.field_survey_id)}></Icon>
+              </DataTable.Cell>
+            </DataTable.Row>)
+            counter += 1
+            to_local_data.push({
+              field_survey_id: value.field_survey_id,
+              local_storage_id: counter,
+              sync_status: 3,
+              mat_characterization: value.mat_characterization,
+              mechanism: value.mechanism,
+              exposure: value.exposure,
+              note: value.note,
+              date: value.date,
+            });
+          }
+          Storage.removeItem("FieldSurveyLogs")
+          Storage.setItem("FieldSurveyLogs", to_local_data)
+        } else {
           field_logs.push(<DataTable.Row style={{ width: 500 }}>
-            <DataTable.Cell style={{ marginRight: -90 }}>{format_date_time["text_format_timestamp"]}</DataTable.Cell>
-            <DataTable.Cell style={{ marginRight: 10 }}>Field Survey Report {value.date.split(' ')[0]}</DataTable.Cell>
-            <DataTable.Cell style={{ marginRight: -190 }}>
-              <Icon name="md-create" style={{ color: "blue" }} onPress={() => this.updateLog(value)}></Icon><Text>   </Text>
-              <Icon name="ios-trash" style={{ color: "red" }} onPress={() => this.removeConfirmation(value.field_survey_id)}></Icon>
-            </DataTable.Cell>
+            <DataTable.Cell style={{ marginRight: 10 }}>No data</DataTable.Cell>
           </DataTable.Row>)
-          counter += 1
-          to_local_data.push({
-            field_survey_id: value.field_survey_id,
-            local_storage_id: counter,
-            sync_status: 3,
-            mat_characterization: value.mat_characterization,
-            mechanism: value.mechanism,
-            exposure: value.exposure,
-            note: value.note,
-            date: value.date,
-          });
         }
-        Storage.removeItem("FieldSurveyLogs")
-        Storage.setItem("FieldSurveyLogs", to_local_data)
-        let data_container = Storage.getItem('FieldSurveyLogs')
-        data_container.then(response => {
-          console.log(response)
-        });
         this.setState({ field_logs: field_logs })
         this.tablePaginate(field_logs)
       })
