+++ app/components/Dashboard/risk_assesment/resources_and_capacities.js
@@ -1,14 +1,14 @@
 import React, { Component } from 'react';
-import { View, Text, TouchableOpacity } from 'react-native';
-import MapSection from './map_section'
-import FamilyRiskProfile from './family_risk_profile'
-import { rassessment_styles } from '../../../assets/styles/risk_assessment_styles'
-import { defaults } from '../../../assets/styles/default_styles'
+import { Text, TouchableOpacity, View } from 'react-native';
 import { ScrollView } from 'react-native-gesture-handler';
-import { DataTable } from 'react-native-paper'
-import Storage from '../../utils/storage'
+import { DataTable } from 'react-native-paper';
 import { NavigationEvents } from 'react-navigation';
-import Notification from '../../utils/alert_notification'
+import { defaults } from '../../../assets/styles/default_styles';
+import { rassessment_styles } from '../../../assets/styles/risk_assessment_styles';
+import Notification from '../../utils/alert_notification';
+import Storage from '../../utils/storage';
+import FamilyRiskProfile from './family_risk_profile';
+import MapSection from './map_section';
 
 export default class ResourcesAndCapacities extends Component {
   constructor(props) {
@@ -67,28 +67,30 @@ export default class ResourcesAndCapacities extends Component {
         let rnc_data = [];
         let to_local_data = [];
         let counter = 0
-        for (const [index, value] of responseJson.entries()) {
+        if (responseJson.length != 0) {
+          for (const [index, value] of responseJson.entries()) {
+            rnc_data.push(<DataTable.Row style={{ width: 500 }}>
+              <DataTable.Cell style={{ marginRight: 10 }}>{value.resource_and_capacity}</DataTable.Cell>
+              <DataTable.Cell style={{ marginRight: 10 }}>{value.status}</DataTable.Cell>
+              <DataTable.Cell style={{ marginRight: 10 }}>{value.owner}</DataTable.Cell>
+            </DataTable.Row>)
+            counter += 1
+            to_local_data.push({
+              resources_and_capacities_id: value.resources_and_capacities_id,
+              local_storage_id: counter,
+              sync_status: 3,
+              resource_and_capacity: value.resource_and_capacity,
+              status: value.status,
+              owner: value.owner
+            })
+          }
+          Storage.removeItem("RiskAssessmentRNC")
+          Storage.setItem("RiskAssessmentRNC", to_local_data)
+        } else {
           rnc_data.push(<DataTable.Row style={{ width: 500 }}>
-            <DataTable.Cell style={{ marginRight: 10 }}>{value.resource_and_capacity}</DataTable.Cell>
-            <DataTable.Cell style={{ marginRight: 10 }}>{value.status}</DataTable.Cell>
-            <DataTable.Cell style={{ marginRight: 10 }}>{value.owner}</DataTable.Cell>
+            <DataTable.Cell style={{ marginRight: 10 }}>No data</DataTable.Cell>
           </DataTable.Row>)
-          counter += 1
-          to_local_data.push({
-            resources_and_capacities_id: value.resources_and_capacities_id,
-            local_storage_id: counter,
-            sync_status: 3,
-            resource_and_capacity: value.resource_and_capacity,
-            status: value.status,
-            owner: value.owner
-          })
         }
-        Storage.removeItem("RiskAssessmentRNC")
-        Storage.setItem("RiskAssessmentRNC", to_local_data)
-        let data_container = Storage.getItem('RiskAssessmentRNC')
-        data_container.then(response => {
-          console.log(response)
-        });
         this.setState({ rnc_data: rnc_data })
         this.tablePaginate(rnc_data)
       })
